Incorporate RequireJS
Configure RequireJS so that third-party modules can be loaded explicitly
(instead of informally referenced on the global scope). Update the
accepted globals to allow references to the AMD functions and disallow
references to the third-party modules.

Wrap the main application code in a call to the AMD `define` function.
This introduces a lot of whitespace change. To see a diff *without* all
the indentation changes, use the following commands:

    $ git show --ignore-all-space ":/Incorporate RequireJS"