Convert to "CommonJS style" AMD
Write the `define` factory function to accept an argument named
`require`, and use it as though it synchronously loaded modules (just
like in CommonJS). This cuts down on the repition in module defintions,
since dependencies do not have to be duplicated in an array and the
factory function's arguments list.